src/pmcfg/mod.rs: moved identify_terminals to src/pmcfg/negra.rs
